# docs/openapi-ext.yml
openapi: 3.0.3
info:
  title: Java Cinema API Challenge - Core Criteria
  description: |-
  version: 0.2.2
servers:
  - url: http://localhost:4000
...
paths:
...
  /movies:
    post:
      tags:
        - movies
      summary: Create a movie
      description: Create a new movie. Optionally, a screenings array can be provided to create screenings when the movie gets created. If no screenings array is provided, the movie should be created as normal.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovie'
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleMovie'
    get:
      tags:
        - movies
      summary: Get all movies
      description: Get a list of every movie.
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllMovies'
  /movies/{id}:
    put:
      tags:
        - movies
      summary: Update a movie
      description: Update an existing movie. If any field is not provided, the original value should not be changed. Any combination of fields can be updated.
      parameters:
        - name: id
          in: path
          description: The ID of the movie to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMovie'
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleMovie'
    delete:
      tags:
        - movies
      summary: Delete a movie
      description: Delete an existing movie. When deleting a movie, all of its screenings should also be deleted.
      parameters:
        - name: id
          in: path
          description: The ID of the movie to delete
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleMovie'
...
components:
  schemas:
...
    CreateMovie:
      type: object
      required: [ title, rating, description, runtimeMins ]
      properties:
        title:
          type: string
          example: Dodgeball
        rating:
          type: string
          example: PG-13
        description:
          type: string
          example: The greatest movie ever made.
        runtimeMins:
          type: integer
          example: 126
        screenings:
          type: array
          items:
            type: object
            required: [ screenNumber, capacity, startsAt ]
            properties:
              screenNumber:
                type: integer
                example: 5
              capacity:
                type: integer
                example: 40
              startsAt:
                type: string
                format: DateTime
                pattern: YYYY-MM-DD HH:MM:SS
                example: 2023-03-19 11:30:00
    UpdateMovie:
      type: object
      properties:
        title:
          type: string
          example: Dodgeball
        rating:
          type: string
          example: PG-13
        description:
          type: string
          example: The greatest movie ever made.
        runtimeMins:
          type: integer
          example: 126
    SingleMovie:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            id:
              type: integer
              example: 1
            title:
              type: string
              example: Dodgeball
            rating:
              type: string
              example: PG-13
            description:
              type: string
              example: The greatest movie ever made.
            runtimeMins:
              type: integer
              example: 126
            createdAt:
              type: string
              format: DateTime
              pattern: YYYY-MM-DDTHH:MM:SS
              example: 2023-03-14T11:01:56.633+00:00
            updatedAt:
              type: string
              format: DateTime
              pattern: YYYY-MM-DDTHH:MM:SS
              example: 2023-03-14T11:01:56.633+00:00
    Movie:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Dodgeball
        rating:
          type: string
          example: PG-13
        description:
          type: string
          example: The greatest movie ever made.
        runtimeMins:
          type: integer
          example: 126
        createdAt:
          type: string
          format: DateTime
          pattern: YYYY-MM-DDTHH:MM:SS
          example: 2023-03-14T11:01:56.633+00:00
        updatedAt:
          type: string
          format: DateTime
          pattern: YYYY-MM-DDTHH:MM:SS
          example: 2023-03-14T11:01:56.633+00:00
    AllMovies:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: array
          items:
            $ref: '#/components/schemas/Movie'
...